Fixed DHCP Proxy Mode Bug
Fixed a bug where if —no-dhcp is specified the DHCP server still starts
because —proxydhcp default is True. Changed the default of —proxydhcp
to False and if —proxydhcp is specified as an argument then it becomes
True and the DHCP server is started in proxy mode.

Added some level of verbosity when each service is started and when
PyPXE as a whole has finished initializing as well as when PyPXE stops
(via keyboard interrupt).

Implemented try/except in __main__

Added minor comments throughout to help understand what’s going on