README.md updates
Removed license from README.md and created a separate LICENSE.md (file
creation done in a separate commit but removal of license from
README.md done in this commit).

Updated README.md to reflect minor change in —proxydhcp argument help
(this was done in a previous commit but noting it here).