'argparse' Module Improvement
If the user does not have the ‘argparse’ module installed then the
server.py will exit and display an error message. Without the module,
server.py will not take in command line arguments properly. This only
affects users using older than Python 2.7 because as of 2.7 it is
maintained in the standard library. The documentation as well as
server.py were updated to reflect this.

Minor whitespace and quote corrections.